[BEAM-5658] Eliminate needless boxing and unboxing

+++ runners/google-cloud-dataflow-java/worker/src/main/java/org/apache/beam/runners/dataflow/worker/WorkItemStatusClient.java
@@ -263,7 +263,7 @@ public class WorkItemStatusClient {
       populateCounterUpdates(status);
     }
 
-    Double throttleTime = extractThrottleTime().doubleValue();
+    double throttleTime = extractThrottleTime();
     status.setTotalThrottlerWaitTimeSeconds(throttleTime);
     return status;
   }
@@ -337,7 +337,7 @@ public class WorkItemStatusClient {
         : executionContext.extractMsecCounters(isFinalUpdate);
   }
 
-  public Long extractThrottleTime() {
+  public long extractThrottleTime() {
     return executionContext == null ? 0L : executionContext.extractThrottleTime();
   }
